Move MenuShell to its own file, and add several missing MenuShell methods
